Scala Stack with jQuery - v0.1 - Proper package names
