+++ src/main/scala/jvm/datastore/entity/ResourceEntity.scala
@@ -1,0 +1,18 @@
+package jvm.datastore.entity
+
+import com.googlecode.objectify.annotation.Entity
+import com.googlecode.objectify.annotation.Cache
+import com.googlecode.objectify.annotation.Id
+import scalajs.shared.Resource
+
+/**
+ * Scala JS can only handle one constructor which needs to be empty to make this class serializable.
+ */
+@Entity
+@Cache
+class ResourceEntity {
+  
+  @Id var id : java.lang.Long = null
+  var r : Resource = null
+  
+}
