+++ src/main/webapp/WEB-INF/web.xml
@@ -9,7 +9,12 @@
     <servlet-class>org.glassfish.jersey.servlet.ServletContainer</servlet-class>
     <init-param>
       <param-name>jersey.config.server.provider.packages</param-name>
-      <param-value>api</param-value>
+      
+      <!-- IMPORTANT -->
+      <!-- Set correct package for your Jersey classes -->
+      <param-value>jvm.api</param-value>
+      
+      
     </init-param>
     <init-param>
       <!-- speed up initial Jersey loading by deactivating WADL -->
@@ -34,7 +39,7 @@
 
   <filter>
     <filter-name>HtmlFilter</filter-name>
-    <filter-class>filter.HtmlFilter</filter-class>
+    <filter-class>jvm.filter.HtmlFilter</filter-class>
   </filter>
   <filter-mapping>
     <filter-name>HtmlFilter</filter-name>
