+++ src/main/webapp/scalajs-source/scalajs/shared/Css.scala
@@ -1,0 +1,12 @@
+package scalajs.shared
+
+import scalatags.Text.all._
+import scalatags.stylesheet.StyleSheet
+
+class Stylisch extends StyleSheet {
+
+  val hidden = cls(display := "none")
+  
+  override def customSheetName = Some("scalatags")
+}
+
