+++ src/main/webapp/scalajs-source/scalajs/shared/html/ServerError.scala
@@ -1,0 +1,20 @@
+package scalajs.shared.html
+
+import scalatags.Text.all._
+import scalatags.Text.tags2.{ title => title2 }
+
+object ServerError {
+  
+  def InternalServerError() = {
+    html(
+        head(
+            title2("5xx")
+            ),
+        body(
+            h1("5xx"),
+            p("Internal server error!")
+            )
+        
+        )
+  }
+}
