--- src/main/webapp/scalajs/shared/Css.scala
@@ -1,12 +1,0 @@
-package shared
-
-import scalatags.Text.all._
-import scalatags.stylesheet.StyleSheet
-
-class Stylisch extends StyleSheet {
-
-  val hidden = cls(display := "none")
-  
-  override def customSheetName = Some("scalatags")
-}
-
