--- src/main/webapp/scalajs/shared/ResourceValidator.scala
@@ -1,6 +1,0 @@
-package shared
-
-object ResourceValidator {
-  
-  def apply(r : Resource)(implicit language : Languages.Language = Languages.default) = if (r.x <= 0) throw new IllegalArgumentException(Translations.resourceValidationError.get(language))
-}
